split network test files
